+++ packages/rules/au/pay/src/index.ts
@@ -4,6 +4,7 @@
  * @since 0.1.0
  */
 export {
+  CalculatePayWithholdings,
   CalculateTakeHomePay,
   TakeHomePayReport,
   TakeHomeScenarioInputSchema,